+++ kernel/arch/x86_64/rtc.c
@@ -4,6 +4,6 @@ void arch_sleep(size_t ns) {
     if (hpet) hpet_sleep(ns);
 }
 
-void uptime(long *sec, long *nsec) {
+void uptime(size_t *sec, size_t *nsec) {
     if (hpet) hpet_read_time(sec, nsec);
 }
